Finish migration from SearchedPhotosFetcher to SearchViewReducer
