+++ FlickrSearch/Modules/Search/ViewModel/SearchViewModel.swift
@@ -10,7 +10,7 @@ final class SearchViewModel {
         (_ viewState: ViewState, _ photos: [Photo]) ->
         BehaviorRelay<(page: SearchPage, viewState: ViewState, newPhotos: [Photo])>
 
-    private let viewStateReducer = SearchViewStateReducer()
+    private let viewStateReducer = SearchViewReducer()
     private var photos = BehaviorSubject(value: [Photo]())
     private var searchPage = SearchPage(query: .empty)
 
@@ -23,8 +23,7 @@ final class SearchViewModel {
     let searchText = BehaviorSubject(value: String.empty)
     let isScrolledToBottom = BehaviorSubject(value: false)
 
-    init(fetcher: FetcherType = Fetcher(apiKey: "3e7cc266ae2b0e0d78e279ce8e361736"),
-         searchPhotosFetcher: SearchedPhotosFetcherType? = nil) {
+    init(fetcher: FetcherType = Fetcher(apiKey: "3e7cc266ae2b0e0d78e279ce8e361736")) {
         self.fetcher = fetcher
 
         searchText.asObserver()
